v1.0.41: Firewall bidirectional forwarding, link-proof relay
- FIREWALL_MODE: link-proof forwarding through link table
- FIREWALL_MODE: reverse hash whitelisting at both forwarding sites
- FIREWALL_MODE: proof firewall exemption from backbone
- FIREWALL_MODE: announce rebroadcast to all interfaces
- FIREWALL_MODE: timestamp-on-load reset for no-NTP case
- Remove DISCOVERY_ADD system in FIREWALL_MODE
- Simplify mode dependencies (no boundary/roaming checks)
- Rename boundary → firewall throughout
- TcpInterface: static frame_buf prevents stack overflow
- TcpInterface: echo prevention via sender skip
- Interface: add isConnected() virtual
- WATCH_LOG system for targeted diagnostics
- Boundary→Firewall rename in logs

+++ Display.h
@@ -940,7 +940,7 @@ void draw_disp_area() {
       // Radio info
       disp_area.setCursor(3, 16);
       if (radio_online) {
-        disp_area.printf("%.3f MHz", (float)lora_freq / 1000000.0);
+        disp_area.printf("%.3f MHz", (double)lora_freq / 1000000.0);
       } else {
         disp_area.print("Radio OFF");
       }
